tests/requirements.txt
playwright==1.43.0
python-dotenv==0.21.0
pytest==7.1.3
pytest-bdd==8.1.0
pytest-playwright==0.4.4
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-base-url==2.0.0
pytest-xdist==3.5.0